requirements.txt
requests
beautifulsoup4
readability-lxml
html2text
lxml_html_clean